requirements.txt
customtkinter>=5.2.0
requests>=2.28.0
pillow>=9.0.0
appdirs>=1.4.4
